apis/requirements.txt
# apis/requirements.txt

fastapi
uvicorn[standard]
pymilvus
psycopg2-binary
sqlalchemy
python-dotenv
kafka-python